调整权限结构

+++ dandelion-upm/dandelion-upm-api/src/main/java/cn/icepear/dandelion/upm/api/domain/entity/SysRoleMenu.java
@@ -23,16 +23,16 @@ public class SysRoleMenu extends Model<SysRoleMenu> {
 	/**
 	 * 角色ID
 	 */
-	private Integer roleId;
+	private Long roleId;
 	/**
 	 * 菜单ID
 	 */
-	private Integer menuId;
+	private Long menuId;
 
 	public SysRoleMenu() {
 	}
 
-	public SysRoleMenu(Integer roleId, Integer menuId) {
+	public SysRoleMenu(Long roleId, Long menuId) {
 		this.roleId = roleId;
 		this.menuId = menuId;
 	}
